Tree Map From Field Measurement

Tree Map From Field Measureme

PRF025_Field

Format

A data frame with one row and 19 variables:

Plot

Character. Plot identifier.

Tree

Integer. Tree identifier within the plot.

Spp

Character. Species code.

Status

Character. Tree status (e.g., L = living, D = dead).

Origin

Character. Tree origin (e.g., N = natural, P = planted).

Dist

Numeric. Distance from plot center to the tree (in meters).

Azim

Numeric. Azimuth from plot center to the tree (in degrees).

DBH

Numeric. Diameter at breast height (in cm).

DbhHt

Numeric. Height at which DBH was measured (in meters).

Ht

Numeric. Tree height (in meters).

LI

Not documented

LO

Not documented

BT

Not documented

Plotsize

Character. Size class of the plot (e.g., "Large").

Dbh_meas

Numeric. Measured DBH (in cm). Redundant with `DBH`?

Field_X

Numeric. Local X coordinate of tree in field measurement system.

Field_Y

Numeric. Local Y coordinate of tree in field measurement system.

Field_Xpj

Numeric. Projected X coordinate (e.g., UTM Easting).

Field_Ypj

Numeric. Projected Y coordinate (e.g., UTM Northing).

Easting

Numeric. Projected X coordinate plot center.

Northing

Numeric. Projected Y coordinate plot center.

Source

Derived from field surveys in forestry plots.

Examples

data(PRF025_Field)
head(PRF025_Field)
#>       Plot Tree Spp Status Origin Dist  Azim  DBH DbhHt   Ht   LI LO   BT
#> 106 PRF025    1  41      L      N  3.8   7.6 14.9   1.3 16.8 <NA> NA <NA>
#> 107 PRF025    2  41      L      N  6.5   5.2 12.1   1.3 17.2 <NA> NA <NA>
#> 108 PRF025    3  41      L      N  8.6   6.8 11.8   1.3 15.0 <NA> NA <NA>
#> 109 PRF025    4  41      L      N  9.2  11.2 12.8   1.3 15.3 <NA> NA <NA>
#> 110 PRF025    5  41      L      N  9.7  16.0 16.7   1.3 16.6 <NA> NA <NA>
#> 111 PRF025    6  41      L      N  0.9 111.1 14.8   1.3 17.6 <NA> NA <NA>
#>     Plotsize Dbh_meas   Field_X    Field_Y Field_Xpj Field_Ypj  Easting
#> 106    Large     14.9 0.5025743  3.7666191  312030.0   5096043 312029.5
#> 107    Large     12.1 0.5891118  6.4732486  312030.1   5096046 312029.5
#> 108    Large     11.8 1.0182741  8.5395034  312030.5   5096048 312029.5
#> 109    Large     12.8 1.7869560  9.0247874  312031.3   5096048 312029.5
#> 110    Large     16.7 2.6736824  9.3242385  312032.2   5096049 312029.5
#> 111    Large     14.8 0.8396582 -0.3239971  312030.4   5096039 312029.5
#>     Northing
#> 106  5096039
#> 107  5096039
#> 108  5096039
#> 109  5096039
#> 110  5096039
#> 111  5096039